+++ threedi_modelchecker/tests/test_checks_timeseries.py
@@ -2,6 +2,8 @@ import pytest
 from threedi_schema import models
 
 from threedi_modelchecker.checks.timeseries import (
+    FirstTimeSeriesEqualTimestepsCheck,
+    TimeSeriesEqualTimestepsCheck,
     TimeseriesExistenceCheck,
     TimeseriesIncreasingCheck,
     TimeseriesRowCheck,
@@ -10,7 +12,72 @@ from threedi_modelchecker.checks.timeseries import (
     TimeseriesValueCheck,
 )
 
-from .factories import BoundaryConditions2DFactory
+from .factories import BoundaryConditions1DFactory, BoundaryConditions2DFactory
+
+
+@pytest.mark.parametrize("check_type", ["1d", "2d"])
+@pytest.mark.parametrize(
+    "timeseries_tuple,expected_invalid",
+    [
+        ((), 0),  # no timeseries
+        (("0,-0.5 \n59,-0.2", "0,-0.5 \n59,-0.2"), 0),  # same levels
+        (
+            ("0,-0.5 \n59,-0.2", "0,-0.5 \n59,-0.3"),
+            0,
+        ),  # same timesteps, different levels
+        (
+            ("0,-0.5 \n59,-0.2", "0,-0.5 \n58,-0.3", "0,-0.5 \n59,-0.3"),
+            1,
+        ),  # differing timestep, one error
+        (
+            ("0,-0.5 \n58,-0.2", "0,-0.5 \n59,-0.3", "0,-0.5 \n59,-0.3"),
+            2,
+        ),  # differing first timestep, all other timeseries error
+    ],
+)
+def test_timeseries_same_timesteps(
+    session, timeseries_tuple, check_type, expected_invalid
+):
+    if check_type == "1d":
+        for timeseries in timeseries_tuple:
+            BoundaryConditions1DFactory(timeseries=timeseries)
+        check = TimeSeriesEqualTimestepsCheck(models.BoundaryCondition1D.timeseries)
+    elif check_type == "2d":
+        for timeseries in timeseries_tuple:
+            BoundaryConditions2DFactory(timeseries=timeseries)
+        check = TimeSeriesEqualTimestepsCheck(models.BoundaryConditions2D.timeseries)
+    invalid = check.get_invalid(session)
+    assert len(invalid) == expected_invalid
+
+
+@pytest.mark.parametrize(
+    "one_d_timeseries_tuple,two_d_timeseries_tuple,expected_invalid",
+    [
+        ((), (), 0),  # no timeseries
+        (("0,-0.5 \n59,-0.2", "0,-0.5 \n59,-0.2"), (), 0),  # no 2d timeseries
+        ((), ("0,-0.5 \n59,-0.2", "0,-0.5 \n59,-0.2"), 0),  # no 1d timeseries
+        (
+            ("0,-0.5 \n59,-0.2", "0,-0.5 \n59,-0.2"),
+            ("0,-0.5 \n59,-0.2", "0,-0.5 \n58,-0.2"),
+            0,
+        ),  # differing second element
+        (
+            ("0,-0.5 \n59,-0.2", "0,-0.5 \n59,-0.2"),
+            ("0,-0.5 \n58,-0.2", "0,-0.5 \n59,-0.2"),
+            1,
+        ),  # differing first element
+    ],
+)
+def test_first_timeseries_same_timesteps(
+    session, one_d_timeseries_tuple, two_d_timeseries_tuple, expected_invalid
+):
+    for timeseries in one_d_timeseries_tuple:
+        BoundaryConditions1DFactory(timeseries=timeseries)
+    for timeseries in two_d_timeseries_tuple:
+        BoundaryConditions2DFactory(timeseries=timeseries)
+    check = FirstTimeSeriesEqualTimestepsCheck()
+    invalid = check.get_invalid(session)
+    assert len(invalid) == expected_invalid
 
 
 @pytest.mark.parametrize("timeseries", ["0,-0.5", "0,-0.5 \n59,-0.5\n60,-0.5\n   "])